requirements.txt
Django==5.0.2
djangorestframework==3.15.1
django-cors-headers==4.4.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
python-decouple==3.8